v1

+++ db/migrate/20180710190116_create_table_notes.rb
@@ -2,7 +2,7 @@ class CreateTableNotes < ActiveRecord::Migration[5.2]
   def change
     create_table :notes do |t|
       t.integer :song_id
-      t.integer :time_in
+      t.float :time_in
       t.float :duration
       t.string :note
     end
